% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2final.R
\name{r2final}
\alias{r2final}
\title{Get R2 values between preference to hear tutor song throuhgout development
and quality of imitation of that song by the adult pupil}
\usage{
r2final()
}
\description{
Get R2 values between preference to hear tutor song throuhgout development
and quality of imitation of that song by the adult pupil
}
\keyword{conditioning,}
\keyword{development}
\keyword{operant}
\keyword{song}
